requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pyjwt==2.9.0
pydantic==2.9.2
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
httpx==0.27.2
telethon==1.36.0
python-dotenv==1.0.1
